+++ src/syntax/italic.ts
@@ -1788,7 +1788,8 @@ export function getItalicSyntax(palette: Palette, italicComments: boolean) {
     // Clojure{{{
     {
       name: "Clojure green",
-      scope: "constant.keyword.clojure",
+      scope:
+        "constant.keyword.clojure, support.variable.clojure, meta.definition.variable.clojure",
       settings: {
         foreground: palette.green,
       },